refac(zk): move `Assembly` to `zk/plonk/keys`
Since `Assembly` is only used to generate keys, this commit moves the
file.
