+++ tachyon/zk/plonk/circuit/BUILD.bazel
@@ -3,18 +3,6 @@ load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
 package(default_visibility = ["//visibility:public"])
 
 tachyon_cc_library(
-    name = "assembly",
-    hdrs = ["assembly.h"],
-    deps = [
-        ":assignment",
-        "//tachyon/base:logging",
-        "//tachyon/base:range",
-        "//tachyon/math/base:rational_field",
-        "//tachyon/zk/plonk/permutation:permutation_assembly",
-    ],
-)
-
-tachyon_cc_library(
     name = "assigned_cell",
     hdrs = ["assigned_cell.h"],
     deps = [
